+++ unite.h
@@ -12,7 +12,7 @@ public:
     bool isTeam();
     bool isAttackable();
     void setAttackable(bool set);
-    void receiveDamage(double dmg);
+    void receiveDamage(int dmg);
     int getVie();
     void setVie(int impact);
     char getTypeMovement(); // Si fonctions redéfinies dans les classes filles -> ajouté virtual ici
